requirements.txt
absl-py==0.4.1
airflow==1.8.0
astor==0.7.1
astroid==2.0.4
backcall==0.1.0
bleach==3.0.0
Click==7.0
cloudpickle==0.5.5
confluent-kafka==0.11.6
cycler==0.10.0
dask==0.19.1
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.2.3
Flask==1.0.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gast==0.2.0
grpcio==1.14.2
h5py==2.8.0
imgaug==0.2.6
ipykernel==5.0.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==0.24
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kafka==1.3.5
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.1
notebook==5.7.0
numpy==1.15.2
opencv-python==3.4.3.18
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.2.0
prometheus-client==0.4.0
prompt-toolkit==2.0.5
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.2.0
pylint==2.1.1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==1.0.0
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.1
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorflow-serving-api==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.10.11
WTForms==2.2.1
